# web_server/docs/swagger.yaml
definitions:
...
  controllers.setRoleReq:
    properties:
      role:
        type: string
    required:
    - role
    type: object
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
  description: 基于 Vue3+UniApp+Gin+Gorm 的社团管理系统后端接口
  title: 高校社团信息管理系统 API
  version: "1.0"
paths:
...
  /leader/clubs/{clubId}/members/{userId}/role:
    post:
      consumes:
      - application/json
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 角色值
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controllers.setRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 负责人设定成员社团内角色
      tags:
      - 负责人
...
  /leader/clubs/{clubId}/users:
    get:
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 获取社团负责人列表
      tags:
      - 负责人
  /leader/users/{userId}/clubs:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - Bearer: []
      summary: 获取用户负责的社团列表
      tags:
      - 负责人
...
  /public/activities:
    get:
      parameters:
      - description: 社团ID
        in: query
        name: clubId
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 开始时间
        in: query
        name: start
        type: string
      - description: 结束时间
        in: query
        name: end
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 活动列表（公开）
      tags:
      - 公共
  /public/announcements:
    get:
      parameters:
      - description: 社团ID
        in: query
        name: clubId
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 公告列表（公开）
      tags:
      - 公共
...
  /public/clubs/{clubId}:
    get:
      parameters:
      - description: 社团ID
        in: path
        name: clubId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 社团详情（公开）
      tags:
      - 公共
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"